+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - response object access modified
 - OauthTokenResponse tests
+- Oauth token refresh request body
 ### Changed
 ### Deprecated
 ### Removed